+++ CHANGES.md
@@ -3,8 +3,9 @@ E-Maj - Change log
 <devel>
 ------
 ###Enhancements:###
-  * Add an optional p_comment parameter to the emaj_create_group() function to
-    directly record a comment for the created group.
+  * Add an optional p_comment parameter to the emaj_create_group(),
+    emaj_set_mark_group() and emaj_set_mark_groups() functions to directly
+    record a comment for the created group or mark.
   * Rollback monitor clients: allow infinite loop with command option -n 0.
   * Both php CLI clients emajRollbackMonitor.php and emajParallelRollback.php
     are deprecated. Use perl clients instead.